Experiment works!!! :-)
We obtain a good estimation of position and velocity.
The motion model includes very small acceleration instead of the acceleration subjected by the slanted surface.
So to finalize, this commit doesn't estimate motion on a slanted surface. (i.e. this estimate a motion on a horizontal surface w/o gravity affecting it, can obtain small accelerations).